Fixed bug that causes app to crash on subsequent launches.

+++ app/src/main/java/com/example/a53914/sievemobileapplication/HomePage.java
@@ -47,7 +47,7 @@ public class HomePage extends AppCompatActivity {
         if(bd != null)
         {
             Boolean delete = (Boolean) bd.get("delete");
-            if (delete){
+            if (delete && delete != null){
                 taskDatabase.taskDao().delete(mTask);
             }
         }
